requirements.txt
alembic==1.4.2
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.11.1
asn1crypto==0.24.0
attrs==19.3.0
Automat==20.2.0
awscli==1.18.62
Babel==2.8.0
blinker==1.4
boto3==1.13.12
botocore==1.16.12
cached-property==1.5.1
cattrs==0.9.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloud-init==19.4
colorama==0.4.3
colorlog==4.0.2
command-not-found==0.3
configobj==5.0.6
configparser==3.5.3
constantly==15.1.0
croniter==0.3.31
cryptography==2.9.2
cssselect==1.1.0
defusedxml==0.6.0
dill==0.3.1.1
distro-info===0.18ubuntu0.18.04.1
dnspython==1.16.0
docutils==0.16
ec2-hibinit-agent==1.0.0
email-validator==1.1.0
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.14
gunicorn==19.10.0
hibagent==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.10.0
json-merge-patch==0.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.1
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.0
netifaces==0.10.4
numpy==1.18.4
oauthlib==2.0.6
PAM==0.4.2
pandas==0.25.3
parsel==1.6.0
pendulum==1.4.4
prison==0.1.3
Protego==0.1.16
psutil==5.7.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.6.1
pygobject==3.26.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyrsistent==0.16.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-daemon==2.1.2
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
pyxdg==0.25
PyYAML==5.3.1
queuelib==1.5.0
requests==2.23.0
requests-unixsocket==0.1.5
rsa==3.4.2
s3transfer==0.3.3
Scrapy==2.1.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.17
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.5
ssh-import-id==5.7
systemd-python==234
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
Twisted==20.3.0
typing==3.7.4.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.25.9
w3lib==1.22.0
Werkzeug==0.16.1
WTForms==2.3.1
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.1.0
